audio/SDL: fix buffer sizing issues causing audio cuts/glitches when CPU is struggling - Use SDL_InitSubSystem(SDL_INIT_AUDIO) instead of SDL_Init() to avoid   reinitializing unrelated SDL subsystems - Double the hardware buffer size (SampleLength * 2) to reduce backend   underruns without affecting emulator buffering logic - Remove the hardcoded 6000-frame fill guard in DoPlay(); the guard was   an arbitrary threshold that caused audio starvation on slow frames.   SDL's internal queue handles backpressure naturally - Read actual channel count from SDL's 'have' spec instead of assuming   the requested value was honored - Minor: NULL → nullptr, coding style cleanup, improved log format

+++ include/ship/audio/SDLAudioPlayer.h
@@ -50,6 +50,6 @@ class SDLAudioPlayer final : public AudioPlayer {
 
   private:
     SDL_AudioDeviceID mDevice = 0; ///< Handle to the opened SDL audio device.
-    int32_t mNumChannels = 2;      ///< Number of output channels (2 for stereo, 6 for 5.1).
+    uint8_t mNumChannels = 2;      ///< Number of output channels (2 for stereo, 6 for 5.1).
 };
 } // namespace Ship
